+++ fbw-common/src/wasm/systems/systems/src/simulation/update_context.rs
@@ -15,7 +15,7 @@ use uom::si::{
 use super::{Read, SimulatorReader};
 use crate::{
     shared::{low_pass_filter::LowPassFilter, MachNumber},
-    simulation::{InitContext, VariableIdentifier},
+    simulation::{InitContext, SideControlling, VariableIdentifier},
 };
 use nalgebra::{Rotation3, Vector3};
 
@@ -238,6 +238,7 @@ pub struct UpdateContext {
     rotation_vel_y_id: VariableIdentifier,
     rotation_vel_z_id: VariableIdentifier,
     aircraft_preset_quick_mode_id: VariableIdentifier,
+    side_controlling_id: VariableIdentifier,
 
     delta: Delta,
     simulation_time: f64,
@@ -279,6 +280,8 @@ pub struct UpdateContext {
 
     rotation_accel: Vector3<AngularAcceleration>,
     rotation_vel: Vector3<AngularVelocity>,
+
+    side_controlling: SideControlling,
 
     /// This is set by the Aircraft Presets to facilitate quick startup or shutdown of the aircraft.
     /// In the context of the apu this means quick startup or shutdown of the apu, and no cooldown
@@ -328,6 +331,7 @@ impl UpdateContext {
     pub(crate) const ROTATION_VEL_Y_KEY: &'static str = "ROTATION VELOCITY BODY Y";
     pub(crate) const ROTATION_VEL_Z_KEY: &'static str = "ROTATION VELOCITY BODY Z";
     pub(crate) const AIRCRAFT_PRESET_QUICK_MODE_KEY: &'static str = "AIRCRAFT_PRESET_QUICK_MODE";
+    pub(crate) const SIDE_CONTROLLING: &'static str = "FBW_PILOT_SEAT";
 
     // Plane accelerations can become crazy with msfs collision handling.
     // Having such filtering limits high frequencies transients in accelerations used for physics
@@ -357,6 +361,7 @@ impl UpdateContext {
         bank: Angle,
         mach_number: MachNumber,
         latitude: Angle,
+        side_controlling: SideControlling,
     ) -> UpdateContext {
         UpdateContext {
             is_ready_id: context.get_identifier(Self::IS_READY_KEY.to_owned()),
@@ -403,6 +408,8 @@ impl UpdateContext {
             rotation_vel_x_id: context.get_identifier(Self::ROTATION_VEL_X_KEY.to_owned()),
             rotation_vel_y_id: context.get_identifier(Self::ROTATION_VEL_Y_KEY.to_owned()),
             rotation_vel_z_id: context.get_identifier(Self::ROTATION_VEL_Z_KEY.to_owned()),
+
+            side_controlling_id: context.get_identifier(Self::SIDE_CONTROLLING.to_owned()),
 
             aircraft_preset_quick_mode_id: context
                 .get_identifier(Self::AIRCRAFT_PRESET_QUICK_MODE_KEY.to_owned()),
@@ -464,6 +471,8 @@ impl UpdateContext {
             rotation_vel: Vector3::default(),
 
             aircraft_preset_quick_mode: false,
+
+            side_controlling,
         }
     }
 
@@ -512,6 +521,8 @@ impl UpdateContext {
 
             aircraft_preset_quick_mode_id: context
                 .get_identifier(Self::AIRCRAFT_PRESET_QUICK_MODE_KEY.to_owned()),
+
+            side_controlling_id: context.get_identifier("SIDE_CONTROLLING".to_owned()),
 
             delta: Default::default(),
             simulation_time: Default::default(),
@@ -566,6 +577,8 @@ impl UpdateContext {
             rotation_vel: Vector3::default(),
 
             aircraft_preset_quick_mode: false,
+
+            side_controlling: SideControlling::CAPTAIN,
         }
     }
 
@@ -661,6 +674,8 @@ impl UpdateContext {
         self.update_relative_wind();
 
         self.update_local_acceleration_plane_reference(delta);
+
+        self.side_controlling = reader.read(&self.side_controlling_id);
     }
 
     // Computes local acceleration including world gravity and plane acceleration
@@ -845,6 +860,10 @@ impl UpdateContext {
 
     pub fn mach_number(&self) -> MachNumber {
         self.mach_number
+    }
+
+    pub fn side_controlling(&self) -> SideControlling {
+        self.side_controlling
     }
 
     pub fn with_delta(&self, delta: Duration) -> Self {